Split out receivePosts
